face-search/requirements.txt
opencv-python
insightface
numpy
psycopg[binary]
pgvector
python-dotenv
onnxruntime
